[HttpFoundation] Fix UploadedFile::getRealPath() if uploading error
If getting an uploaded file through $request->files which had an uploading error
(any of PHP's UPLOAD_ERR_* error except UPLOAD_ERR_NO_FILE), the returned value
of UploadedFile::getRealPath is the absolute path of the public/ directory.
This is inconsistent with parent class behavior: SplFileInfo::getRealPath should
return false if the file does not exist. This could be risky if this value is
used directly, without appropriate checkings on the file.

This overrides SplFileInfo::getRealPath() to return false when the file is invalid.

Fix: #59931